Completed the implementation of shared interaction between data plots and the map display.     Any data point selected on a plot will highlight relevant elements on the map (events/stations) and vice-versa.     E.g. clicking an event on the Shape tab will highlight all the station records for that event at the presently selected frequency band in the visible plot.     This eventing is context sensitive to the tab actively displaying map data and will attempt to highlight the most relevant infromation given the particular kind of data you are viewing.
Implemented the multiple-waveform display. This feature now allows for clicking any data point on the map or any plot associated with 2-N records and getting a display with a paginated and sorted list of all relevent data.
A number of significant performance and stability improvements have been implemented throughout the tool. Most visibly the speed of loading data and running the path correction algorithm should be considerably faster.
The map now supports abritrary user-defined lists of WMS 1.3.0 compliant map tile servers/layers for display on the map. This enables custom map overlays and terrain/overhead imagery/survey data to be displayed for the region you want to calibrate a model for.
The map now automatically displays context sensitive information based on which tab you are actively looking at. For example: the Site tab will display relevant stations/events based on which calibration spectra you are looking at.
Added an optional layer to the map to display station to event paths in the Path view.
A small improvement to the auto-picker to use a slightly more robust SNR calculation given input waveforms with sufficient amounts of noise.
Replaced the Plus icon style with a Hexagonal one to improve readability of plots.
The numbers displayed on the Path hovers are now the number of events (N) rather than number of waveforms (N*Stations).