CCT-1.0.4-M4
With this release we've added a basic capability to run Mw measurements against a calibrated model on new events has been added in support of issue #1.

It's still quiet rudimentary but you can now run the measurements against already loaded data in the database or provide your own data as part of the request to the REST service.

For examples on how to use the service this way we have a Jupyter notebook available in example-notebooks/measure-mws/.