New uncertainty quantification estimates for the corner frequency now available in both CCT and CERT spectral plots. The plot will now show the most likely estimate alone with an approximation of the 95% interval of misfit. The CERT plot also includes uncertainty from both the joint and pair-wise inversion for the spectral ratios. You can also turn the corner estimates on and off like the other plot elements from the legend now.
Added new beta version of export and import functionality for CERT. This allows you to manually adjust the spectral ratio windows for envelopes and then export the project for later use. Output is in a simple JSON document with one JSON object per-line so that the entire document need not be serialized. The format may change significantly based on community feedback while in the beta period. When importing the JSON into the tool it will attempt to map the values to any loaded waveforms based on their metadata and warn the user if it can't find mappings that are valid for any values.

Added a feature to allow for calculating the distances between any specific event in CERT. Right click an event in the events column of the main CERT display and select calculate distance to populate the distance column for all events based on the selected event.

Added additional map highlighting for the CERT map when selecting values in the spectral display and in the event table.

Manual updates to address some outdated instructions based on community feedback.
Re-formulated the observed energy calculation to replace a numerically derived constant to instead use an analytical formula that will take into account the MDAC model's propagation and density parameters. This should allow advanced users more control over the estimates and allowed better support for more types of geology overall.

Fixed a spectra plot replotting issue so that when the window within the ratio detail waveform plots is adjusted, the update will preserve the inversion data and replot it all correctly.